test: split filelist coverage
