+++ test/conftest.py
@@ -17,7 +17,7 @@ def _load_complete_ics55_pdk_available():
 
 complete_ics55_pdk_available = _load_complete_ics55_pdk_available()
 
-FILELIST_INTEGRATION_PREFIX = "test/test_filelist.py::TestCreateWorkspaceIntegration"
+FILELIST_INTEGRATION_PREFIX = "test/data/test_workspace_filelist.py::TestCreateWorkspaceIntegration"
 
 PDK_REQUIRED_TESTS = {
     f"{FILELIST_INTEGRATION_PREFIX}::test_workspace_with_filelist": "",
